Abstract template as separate layer.
