+++ chatbot-cli/src/main/java/com/dopsun/chatbot/cli/sml/TemplateVariablePart.java
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2018 Dop Sun. All rights reserved.
+ */
+
+package com.dopsun.chatbot.cli.sml;
+
+/**
+ * @author Dop Sun
+ * @since 1.0.0
+ */
+final class TemplateVariablePart extends TemplatePart {
+    /**
+     * @param name
+     */
+    public TemplateVariablePart(String name) {
+        super(name);
+    }
+}
