+++ chatbot-cli/src/main/java/com/dopsun/chatbot/cli/sml/WordAndLocation.java
@@ -4,6 +4,8 @@
 
 package com.dopsun.chatbot.cli.sml;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 /**
@@ -11,6 +13,45 @@ import java.util.Objects;
  * @since 1.0.0
  */
 final class WordAndLocation {
+    /**
+     * @param sentence
+     * @return
+     */
+    public static List<WordAndLocation> splitSentence(String sentence) {
+        Objects.requireNonNull(sentence);
+
+        List<WordAndLocation> list = new ArrayList<>();
+
+        String temp = sentence.toLowerCase();
+        int wordStart = -1;
+        for (int index = 0; index < temp.length(); index++) {
+            char c = temp.charAt(index);
+            if (wordStart < 0) {
+                if (Character.isWhitespace(c)) {
+                    continue;
+                }
+
+                wordStart = index;
+            } else {
+                if (Character.isWhitespace(c)) {
+                    StartAndLength location = new StartAndLength(wordStart, index - wordStart);
+                    String word = temp.substring(wordStart, index);
+                    list.add(new WordAndLocation(word, location));
+
+                    wordStart = -1;
+                }
+            }
+        }
+
+        if (wordStart >= 0) {
+            StartAndLength location = new StartAndLength(wordStart, temp.length() - wordStart);
+            String word = temp.substring(wordStart);
+            list.add(new WordAndLocation(word, location));
+        }
+
+        return list;
+    }
+
     private final String word;
     private final StartAndLength location;
